cluster info refactoring
